+++ frontend/src/pages/Catalogue/GridDisplay/components/ItemCard.tsx
@@ -1,6 +1,6 @@
 import * as Accordion from "@radix-ui/react-accordion";
 import { ChevronDownIcon, ExternalLinkIcon } from "@radix-ui/react-icons";
-import { ForageObj } from "../../../types/types";
+import { ForageObj } from "../../../../types/types";
 
 type Props = {
     itemObj: ForageObj;
@@ -10,9 +10,9 @@ const ItemCard = ({ itemObj }: Props) => {
     return (
         <li className="rounded-md border p-4">
             <div className="flex gap-2">
-                <div>
+                {/* <div>
                     <img src={imgUrl} alt={name} />
-                </div>
+                </div> */}
                 <h2>{name}</h2>
             </div>
             <div>
